pyproject.toml
[tool.poetry]
name = "multi-agents"
version = "0.1.0"
description = "Multi Agent System for Support Analysis"
authors = ["Ataliba Miguel <ataliba.miguel@valonylabs.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
streamlit = "^1.31.0"
crewai = "^0.75.1"
crewai-tools = "^0.13.2"
plotly = "^5.18.0"
pandas = "^2.1.4"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
asyncio = "^3.4.3"
aiohttp = "^3.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"